% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git.user.R
\name{git.get.user}
\alias{git.get.user}
\title{Get git user information}
\usage{
git.get.user(global = FALSE)
}
\arguments{
\item{global}{if TRUE, retrieves system-global git configuration (i.e. --global);
by default, retrieves configuration as used by the current git repository.}
}
\description{
\code{git.get.user} retrieves git configuration describing the default user.
}
\note{
it requires git to be installed on the system.

only tested on Windows as of now
}

