% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git.user.R
\name{git.set.user}
\alias{git.set.user}
\title{Set git user information}
\usage{
git.set.user(name, email, global = FALSE)
}
\arguments{
\item{name}{a character string containing full name of the user; e.g. "John Smith"}

\item{email}{a character string containing email of the user; e.g. "jsmith@example.com"}

\item{global}{if TRUE, changes the system-global configuration of git; by default, changes
the configuration local to the current git repository.}
}
\description{
\code{git.set.user} sets the user name and email to be used by git.
}

