\name{git.user-package}
\alias{git.user-package}
\alias{git.user}
\docType{package}
\title{
Easily change git username config
}
\description{
Allows easy changing of "user" config settings for git version control (git config user.name, user.email).
}
\details{
\tabular{ll}{
Package: \tab git.user\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-04-30\cr
License: \tab MIT\cr
}

}
\author{
Mateusz Czapliński

Maintainer: Mateusz Czapliński <czapkofan@gmail.com>
}
\references{
}
\keyword{ package }
\seealso{
\code{\link[git2r:git2r-package]{git2r}}
\code{\link[installr:installr-package]{installr}}
}
